refactor: isolate WalletConnect lifecycle state
