+++ apps/mobile/src/core/walletconnect/requestBridge.ts
@@ -18,9 +18,9 @@ import {
   getWalletConnectSession,
   getWalletConnectSessionOrigin,
   isWalletConnectMethodApproved,
-  resolveWalletConnectAccount,
   syncWalletConnectSessionsFromClient,
 } from './sessions';
+import { resolveWalletConnectAccount } from './sessionAccountResolution';
 import {
   WALLETCONNECT_READ_ONLY_RPC_METHODS,
   WALLETCONNECT_SIGN_METHODS,
